ScaledTime NumberGenerator now always casts time values to float

+++ numbergen/__init__.py
@@ -390,7 +390,7 @@ class ScaledTime(NumberGenerator, TimeDependent):
 
 
     def __call__(self):
-        return self.time_fn() * self.factor
+        return float(self.time_fn() * self.factor)
